requirements.txt
Django>=5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt>=5.3.0
django-cors-headers==4.4.0
django-filter==24.3
drf-spectacular>=0.27.0
psycopg2-binary>=2.9.9
mysqlclient>=2.2.0
pinecone>=5.0.0
langgraph>=0.2.12
langchain-openai>=0.2.2
PyPDF2>=3.0.1
python-docx>=1.1.2
beautifulsoup4>=4.12.0
openai>=1.0.0
python-decouple>=3.8
requests>=2.31.0
spacy>=3.7.0

